requirements.txt
algorithmia
python-telegram-bot
